backend_audio/requirements.txt
google-cloud-aiplatform
google-cloud-speech
pyaudio
